Move Metal renderer to shared MetalRenderer folder
