+++ iOS/GameViewController.h
@@ -8,7 +8,7 @@
 #import <UIKit/UIKit.h>
 #import <Metal/Metal.h>
 #import <MetalKit/MetalKit.h>
-#import "../macOS/Renderer/MetalRenderer.h"
+#import "../MetalRenderer/MetalRenderer.h"
 
 // Our iOS view controller
 @interface GameViewController : UIViewController
