+++ macOS/GameViewController.h
@@ -8,7 +8,7 @@
 #import <Cocoa/Cocoa.h>
 #import <Metal/Metal.h>
 #import <MetalKit/MetalKit.h>
-#import "Renderer/MetalRenderer.h"
+#import "../MetalRenderer/MetalRenderer.h"
 
 @interface GameViewController : NSViewController
