+++ verify_structure.sh
@@ -23,10 +23,10 @@ files=(
     "macOS/main.m"
     "macOS/GameViewController.h"
     "macOS/GameViewController.mm"
-    "macOS/Renderer/MetalRenderer.h"
-    "macOS/Renderer/MetalRenderer.mm"
-    "macOS/Renderer/Shaders.metal"
-    "macOS/Renderer/MetalShaderTypes.h"
+    "MetalRenderer/MetalRenderer.h"
+    "MetalRenderer/MetalRenderer.mm"
+    "MetalRenderer/Shaders.metal"
+    "MetalRenderer/MetalShaderTypes.h"
     "macOS/Info.plist"
     "iOS/main.m"
     "iOS/AppDelegate.h"
